+++ lib/api/api_connect.dart
@@ -3,18 +3,18 @@ import 'package:http/http.dart' as http;
 import 'dart:async';
 import 'dart:convert';
 import 'package:Solon/api/message.dart';
-import 'package:Solon/api/user.dart';
+// import 'package:Solon/api/user.dart';
 import 'package:Solon/api/proposal.dart';
-import 'package:Solon/api/comment.dart';
-import 'package:Solon/api/event.dart';
-import 'package:Solon/api/forumpost.dart';
-import 'package:Solon/api/vote.dart';
+// import 'package:Solon/api/comment.dart';
+// import 'package:Solon/api/event.dart';
+// import 'package:Solon/api/forumpost.dart';
+// import 'package:Solon/api/vote.dart';
 import 'package:Solon/api/register.dart';
 import 'package:Solon/api/login.dart';
-import 'package:Solon/api/encryptPassword.dart';
+// import 'package:Solon/api/encryptPassword.dart';
 import 'dart:io';
 import 'package:flutter/services.dart';
-import 'package:intl/intl.dart';
+// import 'package:intl/intl.dart';
 // import "package:pointycastle/export.dart";
 // import 'package:flutter_string_encryption/flutter_string_encryption.dart';
 
@@ -176,10 +176,10 @@ class APIConnect {
     return json.decode(response.body);
   }
 
-  static Future<Vote> connectVotes() async {
-    final response = await http.get(
-      "$_url/votes",
-      headers: {HttpHeaders.authorizationHeader: await loadHeader()},
-    );
-  }
+  // static Future<Vote> connectVotes() async {
+  //   final response = await http.get(
+  //     "$_url/votes",
+  //     headers: {HttpHeaders.authorizationHeader: await loadHeader()},
+  //   );
+  // }
 }