+++ lib/auth/sign_up.dart
@@ -1,4 +1,4 @@
-import 'dart:convert';
+// import 'dart:convert';
 
 import 'package:Solon/api/api_connect.dart';
 import 'package:Solon/app_localizations.dart';